Upgrade to Pulumi v2.12.0 and pulumi-terraform-bridge v2.11.0

+++ sdk/python/pulumi_mongodbatlas/maintenance_window.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 
 __all__ = ['MaintenanceWindow']
@@ -15,10 +15,10 @@ class MaintenanceWindow(pulumi.CustomResource):
     def __init__(__self__,
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
-                 day_of_week: Optional[pulumi.Input[float]] = None,
+                 day_of_week: Optional[pulumi.Input[int]] = None,
                  defer: Optional[pulumi.Input[bool]] = None,
-                 hour_of_day: Optional[pulumi.Input[float]] = None,
-                 number_of_deferrals: Optional[pulumi.Input[float]] = None,
+                 hour_of_day: Optional[pulumi.Input[int]] = None,
+                 number_of_deferrals: Optional[pulumi.Input[int]] = None,
                  project_id: Optional[pulumi.Input[str]] = None,
                  __props__=None,
                  __name__=None,
@@ -59,10 +59,10 @@ class MaintenanceWindow(pulumi.CustomResource):
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] day_of_week: Day of the week when you would like the maintenance window to start as a 1-based integer: S=1, M=2, T=3, W=4, T=5, F=6, S=7.
+        :param pulumi.Input[int] day_of_week: Day of the week when you would like the maintenance window to start as a 1-based integer: S=1, M=2, T=3, W=4, T=5, F=6, S=7.
         :param pulumi.Input[bool] defer: Defer maintenance for the given project for one week.
-        :param pulumi.Input[float] hour_of_day: Hour of the day when you would like the maintenance window to start. This parameter uses the 24-hour clock, where midnight is 0, noon is 12 (Time zone is UTC).
-        :param pulumi.Input[float] number_of_deferrals: Number of times the current maintenance event for this project has been deferred, you can set a maximum of 2 deferrals.
+        :param pulumi.Input[int] hour_of_day: Hour of the day when you would like the maintenance window to start. This parameter uses the 24-hour clock, where midnight is 0, noon is 12 (Time zone is UTC).
+        :param pulumi.Input[int] number_of_deferrals: Number of times the current maintenance event for this project has been deferred, you can set a maximum of 2 deferrals.
         :param pulumi.Input[str] project_id: The unique identifier of the project for the Maintenance Window.
         """
         if __name__ is not None:
@@ -100,10 +100,10 @@ class MaintenanceWindow(pulumi.CustomResource):
     def get(resource_name: str,
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
-            day_of_week: Optional[pulumi.Input[float]] = None,
+            day_of_week: Optional[pulumi.Input[int]] = None,
             defer: Optional[pulumi.Input[bool]] = None,
-            hour_of_day: Optional[pulumi.Input[float]] = None,
-            number_of_deferrals: Optional[pulumi.Input[float]] = None,
+            hour_of_day: Optional[pulumi.Input[int]] = None,
+            number_of_deferrals: Optional[pulumi.Input[int]] = None,
             project_id: Optional[pulumi.Input[str]] = None,
             start_asap: Optional[pulumi.Input[bool]] = None) -> 'MaintenanceWindow':
         """
@@ -113,10 +113,10 @@ class MaintenanceWindow(pulumi.CustomResource):
         :param str resource_name: The unique name of the resulting resource.
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] day_of_week: Day of the week when you would like the maintenance window to start as a 1-based integer: S=1, M=2, T=3, W=4, T=5, F=6, S=7.
+        :param pulumi.Input[int] day_of_week: Day of the week when you would like the maintenance window to start as a 1-based integer: S=1, M=2, T=3, W=4, T=5, F=6, S=7.
         :param pulumi.Input[bool] defer: Defer maintenance for the given project for one week.
-        :param pulumi.Input[float] hour_of_day: Hour of the day when you would like the maintenance window to start. This parameter uses the 24-hour clock, where midnight is 0, noon is 12 (Time zone is UTC).
-        :param pulumi.Input[float] number_of_deferrals: Number of times the current maintenance event for this project has been deferred, you can set a maximum of 2 deferrals.
+        :param pulumi.Input[int] hour_of_day: Hour of the day when you would like the maintenance window to start. This parameter uses the 24-hour clock, where midnight is 0, noon is 12 (Time zone is UTC).
+        :param pulumi.Input[int] number_of_deferrals: Number of times the current maintenance event for this project has been deferred, you can set a maximum of 2 deferrals.
         :param pulumi.Input[str] project_id: The unique identifier of the project for the Maintenance Window.
         :param pulumi.Input[bool] start_asap: Flag indicating whether project maintenance has been directed to start immediately. If you request that maintenance begin immediately, this field returns true from the time the request was made until the time the maintenance event completes.
         """
@@ -134,7 +134,7 @@ class MaintenanceWindow(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="dayOfWeek")
-    def day_of_week(self) -> pulumi.Output[float]:
+    def day_of_week(self) -> pulumi.Output[int]:
         """
         Day of the week when you would like the maintenance window to start as a 1-based integer: S=1, M=2, T=3, W=4, T=5, F=6, S=7.
         """
@@ -150,7 +150,7 @@ class MaintenanceWindow(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="hourOfDay")
-    def hour_of_day(self) -> pulumi.Output[float]:
+    def hour_of_day(self) -> pulumi.Output[int]:
         """
         Hour of the day when you would like the maintenance window to start. This parameter uses the 24-hour clock, where midnight is 0, noon is 12 (Time zone is UTC).
         """
@@ -158,7 +158,7 @@ class MaintenanceWindow(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="numberOfDeferrals")
-    def number_of_deferrals(self) -> pulumi.Output[float]:
+    def number_of_deferrals(self) -> pulumi.Output[int]:
         """
         Number of times the current maintenance event for this project has been deferred, you can set a maximum of 2 deferrals.
         """
